Revert "Revert "プロジェクト全体に対してリファクタリングを行いました。""
