+++ .gitignore
@@ -1,4 +1,5 @@
 session_passfile
 sql_passfile
 /node_modules/
+/.ssh/
