+++ bookshelf_app/home.html
@@ -14,7 +14,7 @@
   <script src="https://unpkg.com/react@18/umd/react.production.min.js" crossorigin></script>
   <script src="https://unpkg.com/react-dom@18/umd/react-dom.production.min.js" crossorigin></script>
   <!-- Load our React component. -->
-  <script src="home.js"></script>
+  <script type="module" src="home.js"></script>
 </body>
 
 </html>
