+++ bookshelf_app/home.js
@@ -1,196 +1,10 @@
 'use strict';
 
-const e = React.createElement;
-async function getBookJson(isbn) {
-  const url = "https://api.openbd.jp/v1/get?isbn=" + isbn;
-  try {
-    const response = await fetch(url);
-    const data = await response.json();
-    if (data && data[0]) {
-      console.log("call isbn api success");
-      return data[0];
-    } else {
-      return {
-        summary: {
-          title: "本の情報を取得できませんでした。isbn:" + isbn
-        }
-      };
-    }
-  } catch (error) {
-    console.error("Error occurred while fetching book data:", error);
-    return {
-      summary: {
-        title: "本のデータを取得中にエラーが発生しました isbn:" + isbn
-      }
-    };
-  }
-}
-function BookButton(props) {
-  return /*#__PURE__*/React.createElement("button", {
-    onClick: () => props.onClick(book)
-  }, props.bookButtonText());
-}
-function ShowBooks({
-  books,
-  bookButton
-}) {
-  if (books.length === 0) {
-    return /*#__PURE__*/React.createElement("div", null, "\u672C\u304C\u767B\u9332\u3055\u308C\u3066\u3044\u307E\u305B\u3093\u3002");
-  }
-  const readStateText = read_state => {
-    switch (read_state) {
-      case 0:
-        return "未読";
-      case 1:
-        return "読み止し";
-      case 2:
-        return "読了";
-    }
-  };
-  const listItems = books.map((book, index) => {
-    if (book.detail !== undefined) {
-      return /*#__PURE__*/React.createElement("tbody", {
-        key: book.id
-      }, /*#__PURE__*/React.createElement("tr", {
-        className: "bookDetail"
-      }, /*#__PURE__*/React.createElement("td", null, book.detail.summary.title), /*#__PURE__*/React.createElement("td", null, /*#__PURE__*/React.createElement("img", {
-        src: book.detail.summary.cover,
-        alt: "book_image",
-        width: "100",
-        height: " auto"
-      })), /*#__PURE__*/React.createElement("td", null, ":"), /*#__PURE__*/React.createElement("td", null, readStateText(book.read_state)), /*#__PURE__*/React.createElement("td", null, bookButton(index))));
-    } else {
-      return /*#__PURE__*/React.createElement("tbody", {
-        key: book.id
-      }, /*#__PURE__*/React.createElement("tr", {
-        className: "bookDetails"
-      }, /*#__PURE__*/React.createElement("td", null)));
-    }
-  });
-  return /*#__PURE__*/React.createElement("table", null, listItems);
-}
-class BookShowState extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      index: -1
-    };
-  }
-  ShowBookDetail(bookDetail) {
-    let bookCollateralDetailHtml;
-    if (bookDetail.onix) {
-      bookCollateralDetailHtml = bookDetail.onix.CollateralDetail.TextContent.map((textContent, index) => /*#__PURE__*/React.createElement("p", {
-        key: index
-      }, textContent.Text));
-    }
-    return /*#__PURE__*/React.createElement("div", null, /*#__PURE__*/React.createElement("h2", null, bookDetail.summary.title), /*#__PURE__*/React.createElement("img", {
-      src: bookDetail.summary.cover,
-      alt: "book_image",
-      width: "300",
-      height: " auto"
-    }), /*#__PURE__*/React.createElement("p", null, bookDetail.summary.author), /*#__PURE__*/React.createElement("div", null, bookCollateralDetailHtml), /*#__PURE__*/React.createElement("hr", null));
-  }
-  render() {
-    let {
-      books
-    } = this.props;
-    let bookDetailHtml;
-    const bookButton = index => {
-      return /*#__PURE__*/React.createElement("button", {
-        onClick: () => this.setState({
-          index: index
-        }),
-        disabled: books[index].detail.onix === undefined
-      }, "\u8A73\u7D30\u3092\u898B\u308B");
-    };
-    //詳細を見るボタンが押されたら、その本の詳細を表示　
-    if (this.state.index !== -1) {
-      bookDetailHtml = this.ShowBookDetail(books[this.state.index].detail);
-    }
-    return /*#__PURE__*/React.createElement("div", null, /*#__PURE__*/React.createElement("hr", null), bookDetailHtml, /*#__PURE__*/React.createElement(ShowBooks, {
-      books: books,
-      bookButton: bookButton
-    }));
-  }
-}
-;
-function IsbnInputArea({
-  inputingIsbn,
-  inputOnChange,
-  submitOnClick
-}) {
-  return /*#__PURE__*/React.createElement("div", {
-    className: "form-group"
-  }, /*#__PURE__*/React.createElement("div", null, "ISBN\u30B3\u30FC\u30C9", /*#__PURE__*/React.createElement("input", {
-    type: "text",
-    name: "isbn",
-    value: inputingIsbn,
-    inputMode: "numeric",
-    onChange: inputOnChange
-  }), " "), /*#__PURE__*/React.createElement("button", {
-    type: "submit",
-    onClick: submitOnClick
-  }, "Add"));
-}
-class BookAddState extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      inputingIsbn: ""
-    };
-  }
-  render() {
-    const {
-      submitOnClick,
-      books
-    } = this.props;
-    return /*#__PURE__*/React.createElement("div", null, /*#__PURE__*/React.createElement("hr", null), /*#__PURE__*/React.createElement(IsbnInputArea, {
-      inputingIsbn: this.state.inputingIsbn,
-      submitOnClick: () => {
-        submitOnClick(this.state.inputingIsbn);
-        this.setState({
-          inputingIsbn: ""
-        });
-      },
-      inputOnChange: e => {
-        this.setState({
-          inputingIsbn: e.target.value.replace(/[^0-9]/g, "")
-        });
-      }
-    }), /*#__PURE__*/React.createElement("hr", null), /*#__PURE__*/React.createElement(ShowBooks, {
-      books: books,
-      bookButton: id => ""
-    }));
-  }
-}
-function BookReadingChangeState({
-  books,
-  changeReadState
-}) {
-  const bookButton = index => {
-    return /*#__PURE__*/React.createElement("button", {
-      onClick: () => changeReadState(index)
-    }, "\u8AAD\u66F8\u72B6\u614B\u5909\u66F4");
-  };
-  return /*#__PURE__*/React.createElement("div", null, /*#__PURE__*/React.createElement(ShowBooks, {
-    books: books,
-    bookButton: bookButton
-  }));
-}
-function BookDeleteState({
-  books,
-  deleteBook
-}) {
-  const bookButton = index => {
-    return /*#__PURE__*/React.createElement("button", {
-      onClick: () => deleteBook(index)
-    }, "\u672C\u3092\u524A\u9664");
-  };
-  return /*#__PURE__*/React.createElement("div", null, /*#__PURE__*/React.createElement(ShowBooks, {
-    books: books,
-    bookButton: bookButton
-  }));
-}
+import BookShowState from './BookShowState.js';
+import BookAddState from './BookAddState.js';
+import BookReadingChangeState from './BookReadingChangeState.js';
+import BookDeleteState from './BookDeleteState.js';
+import CallAPIRapper from './CallAPIRapper.js';
 function ModeSelecter(props) {
   return /*#__PURE__*/React.createElement("div", {
     className: "form-group"
@@ -214,7 +28,6 @@ function ModeSelecter(props) {
   }), "\u66F8\u7C4D\u524A\u9664\u30E2\u30FC\u30C9");
 }
 class Bookshelf extends React.Component {
-  ///this class Call API
   constructor(props) {
     super(props);
     this.state = {
@@ -225,23 +38,17 @@ class Bookshelf extends React.Component {
     };
     this.loadIsbn();
   }
-  async loadIsbn() {
+  loadIsbn = async () => {
     try {
-      const response = await fetch(`/api/get_have_books`, {
-        method: 'GET'
-      });
-      const books = await response.json();
-      for (const book of books) {
-        book.detail = await getBookJson(book.isbn);
-      }
+      const books = await CallAPIRapper.loadIsbn();
       this.setState({
         books: books
       });
     } catch (error) {
       console.error(error);
     }
-  }
-  async registerIsbn(inputingIsbn) {
+  };
+  registerIsbn = async inputingIsbn => {
     try {
       if (inputingIsbn.length === 0) {
         this.setState({
@@ -249,18 +56,8 @@ class Bookshelf extends React.Component {
         });
         return;
       }
-      let send_data = {
-        isbn: inputingIsbn
-      };
-      const response = await fetch('/api/register_book', {
-        method: 'POST',
-        headers: {
-          'Accept': 'application/json',
-          'Content-Type': 'application/json'
-        },
-        body: JSON.stringify(send_data)
-      });
-      const json = await response.json();
+      const json = await CallAPIRapper.registerIsbn(inputingIsbn);
+      console.log("res: " + json.text);
       if (json.text === 'success') {
         this.setState({
           server_response: '登録できました！'
@@ -268,7 +65,6 @@ class Bookshelf extends React.Component {
         this.setState({
           inputingIsbn: ''
         });
-        json.book.detail = await getBookJson(json.book.isbn);
         this.setState({
           books: this.state.books.concat([json.book])
         });
@@ -291,23 +87,11 @@ class Bookshelf extends React.Component {
         server_response: 'サーバーエラーが発生しました。'
       });
     }
-  }
-  async changeReadState(index, new_read_state) {
+  };
+  changeReadState = async (index, new_read_state) => {
     try {
-      let send_data = {
-        book: this.state.books[index],
-        new_read_state: new_read_state
-      };
-      const response = await fetch('/api/change_read_state', {
-        method: 'POST',
-        headers: {
-          'Accept': 'application/json',
-          'Content-Type': 'application/json'
-        },
-        body: JSON.stringify(send_data)
-      });
-      const json = await response.json();
-      console.log(json.text);
+      const json = await CallAPIRapper.changeReadState(this.state.books[index], new_read_state);
+      console.log("res: " + json.text);
       if (json.text === 'success') {
         this.setState({
           server_response: '変更できました！'
@@ -337,22 +121,11 @@ class Bookshelf extends React.Component {
         server_response: 'サーバーエラーが発生しました。'
       });
     }
-  }
-  async deleteBook(index) {
+  };
+  deleteBook = async index => {
     try {
-      let send_data = {
-        book: this.state.books[index]
-      };
-      const response = await fetch('/api/delete_book', {
-        method: 'POST',
-        headers: {
-          'Accept': 'application/json',
-          'Content-Type': 'application/json'
-        },
-        body: JSON.stringify(send_data)
-      });
-      const json = await response.json();
-      console.log(json.text);
+      const json = await CallAPIRapper.deleteBook(this.state.books[index]);
+      console.log("res: " + json.text);
       if (json.text === 'success') {
         this.setState({
           server_response: this.state.books[index].detail.summary.title + ' を削除しました。'
@@ -371,20 +144,21 @@ class Bookshelf extends React.Component {
         server_response: 'サーバーエラーが発生しました。'
       });
     }
-  }
-  async shareUrlCopyToCrip() {
+  };
+  shareUrlCopyToCrip = async () => {
     try {
       const response = await fetch(`/api/get_user_id`, {
         method: 'GET'
       });
       const json = await response.json();
-      console.log(json.user_id);
       const shareUrl = `${window.location.origin}/shared_books/${json.user_id}`;
       navigator.clipboard.writeText(shareUrl).then(() => {
+        console.log(shareUrl + " was copyed");
         this.setState({
           server_response: 'クリップボードに共有用URLをコピーしました。'
         });
       }, () => {
+        console.log(shareUrl + " :can not copy to clipboard");
         this.setState({
           server_response: 'URL: ' + shareUrl
         });
@@ -394,7 +168,7 @@ class Bookshelf extends React.Component {
         server_response: 'サーバーエラーが発生しました。'
       });
     }
-  }
+  };
   render() {
     const modeStateHtml = mode_state => {
       switch (mode_state) {
@@ -404,7 +178,7 @@ class Bookshelf extends React.Component {
           });
         case 1:
           return /*#__PURE__*/React.createElement(BookAddState, {
-            submitOnClick: inputedIsbn => this.registerIsbn(inputedIsbn),
+            submitOnClick: this.registerIsbn,
             books: this.state.books
           });
         case 2:
