+++ bookshelf_app/index.html
@@ -14,7 +14,7 @@
 <ul>
 <li>このサイトは頻繁に利用できない状態になります。また、<span style="text-decoration: underline;">予告なしに公開終了</span>する場合があります。</li>
 <li>サービスに登録したデータは、<span style="text-decoration: underline;">予告なしに削除</span>される場合があります。</li>
-<li><span style="text-decoration: underline;">セキュリティ対策</span>をほとんどしていません。<span style="text-decoration: underline;">ユーザー名やパスワードは公開されても問題ないものにしてください</span>。</li>
+<li><span style="text-decoration: underline;">セキュリティ対策</span>が不完全です。<span style="text-decoration: underline;">ユーザー名やパスワードは公開されても問題ないものにしてください</span>。</li>
 <li>上記含め、当該サイトに起因してご利用者様および第三者に損害が発生したとしても、開発者は責任を負わないものとします。<br /><br /></li>
 </ul>
 	<div>
