+++ bookshelf_app/src/login.js
@@ -1,5 +1,4 @@
 'use strict';
-const e = React.createElement;
 class Login extends React.Component {
 
   constructor(props) {
@@ -12,26 +11,30 @@ class Login extends React.Component {
   }
 
   render() {
-    const handleSubmit = (e) => {
+    const handleSubmit = async (e) => {
       let send_data = { name: this.state.name, pass: this.state.password };
-      
-      fetch('/login', {
-        method: 'POST',
-        headers: {
-          'Accept': 'application/json',
-          'Content-Type': 'application/json'
-        },
-        body: JSON.stringify(send_data),
-      })
-        .then((data) => data.json())
-        .then((json) => {
-          if (json.text === 'user or password is wrong') {
-            this.setState({ login_state_text: 'ログインに失敗しました' });
-          } else if (json.text === 'success') {
-            location.href = '/home';
-          }
+      e.preventDefault();
+
+      try {
+        const response = await fetch('/login', {
+          method: 'POST',
+          headers: {
+            'Accept': 'application/json',
+            'Content-Type': 'application/json'
+          },
+          body: JSON.stringify(send_data),
         });
-        e.preventDefault();
+
+        const json = await response.json();
+        console.log("res: " + json.text);
+        if (json.text === 'user or password is wrong') {
+          this.setState({ login_state_text: 'ログインに失敗しました' });
+        } else if (json.text === 'success') {
+          location.href = '/home';
+        }
+      } catch (error) {
+        console.error('エラーが発生しました', error);
+      }
     }
     return (
       <div>
