+++ package.json
@@ -1,21 +1,27 @@
 {
   "name": "cyanobell",
   "version": "1.0.0",
-  "main": "books_server.js",
+  "main": "server.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node books_server.js"
+    "start": "node server.js"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "babel-cli": "^7.21.0",
+    "@babel/cli": "^7.21.0",
+    "@babel/plugin-transform-modules-commonjs": "^7.21.2",
+    "@babel/plugin-transform-runtime": "^7.21.4",
+    "@babel/preset-react": "^7.18.6",
     "babel-preset-react-app": "^3.1.2",
     "bcrypt": "^5.1.0",
     "body-parser": "^1.20.2",
+    "core-js": "^3.30.1",
     "express": "^4.18.2",
     "express-session": "^1.17.3",
-    "mysql": "^2.18.1"
+    "mysql": "^2.18.1",
+    "mysql2": "^3.2.4",
+    "regenerator-runtime": "^0.13.11"
   },
   "keywords": [],
   "description": ""
